WIP refactor util methods
